+++ psi-cpp/psi-stkmanage/dao/Pyrk/PyrkDao.h
@@ -14,14 +14,14 @@ public:
 	// 根据单据编号查其id
 	string selectBillIdByBillNo(const string& billNo);
 	// 插入数据
-	uint64_t insert(const StkIoDO& iObj);
-	uint64_t insert(const StkIoEntryDO& iObj);
+	int insert(const StkIoDO& iObj);
+	int insert(const StkIoEntryDO& iObj);
 	// 上传文件
 	string insertFile(const string& fileName);
 	// 审核单据更新数据
-	uint64_t updateApproval(const StkIoDO& iObj);
+	int updateApproval(const StkIoDO& iObj);
 	// 修改数据
-	uint64_t update(const StkIoDO& iObj);
-	uint64_t update(const StkIoEntryDO& iObj);
+	int update(const StkIoDO& iObj);
+	int update(const StkIoEntryDO& iObj);
 };
 #endif
